히히
프로 시쥬얼이랑 다이얼로그 업뎃 끝

+++ Manzo/Manzo/Engine/Procedual.h
@@ -5,11 +5,12 @@
 
 class ProceduralChain {
 public:
-    void Initialize(int count, int defaultSize);
+    void Initialize(int count, float defaultSize);
     void Update(vec2 headPosition, float followSpeed);
     const std::vector<vec2>& GetPositions() const;
 
 private:
     std::vector<vec2> positions;
-    std::vector<int> circle_size;
+    std::vector<float> circle_size;
 };
+
